fix: resolve gosec G115 and gocritic exitAfterDefer linter warnings

+++ internal/domain/subscription.go
@@ -27,6 +27,6 @@ type SummaryResult struct {
 type ListFilter struct {
 	UserID      *uuid.UUID
 	ServiceName *string
-	Limit       int
-	Offset      int
+	Limit       uint64
+	Offset      uint64
 }
